requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
psutil==6.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
